+++ pos_keyboard/doc/changelog.rst
@@ -1,7 +1,18 @@
-.. _changelog:
+`1.1.3`
+-------
 
-Updates
-=======
+**Fix:** Enter did not work for some pop-ups
+
+`1.1.2`
+-------
+
+**Fix:** Error on clicking numbers and enter in selection popups
+
+
+`1.1.1`
+-------
+
+**Fix:** Incorrect input data on password pop-up if the first clicked button was `backspace`
 
 `1.1.0`
 -------