+++ pos_partner_sync/doc/changelog.rst
@@ -1,3 +1,9 @@
+`2.0.2`
+-------
+
+**Fix:** Client line data updating
+**Fix:** Several updates within the one second were denied by built-in behavior
+
 `2.0.1`
 -------